requirements.txt
openshift
PyYAML
requests
requests_mock
jsonpath-rw
connexion[swagger-ui]
connexion
pydantic
jinja2
escapism
tornado
jupyterhub
pytest
